requirements.txt
requests==2.31.0
wget==3.2
hydra-core==1.3.2
snakemake==8.1.2
selenium==4.29.0
chromedriver-binary==135.0.7030.0.0
tqdm==4.67.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0